Store ClientNumber as int instead of converting back & forth to long

+++ scope.h
@@ -107,7 +107,7 @@ extern char AudioServerHostName[MAXHOSTNAMELEN];
 struct fdinfo
 {
   Boolean Server;
-  long    ClientNumber;
+  int     ClientNumber;
   FD	  pair;
   unsigned char	  *buffer;
   int	  bufcount;
